+++ packages/rest/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @guildedts/rest
 
+## 0.2.0
+
+### Minor Changes
+
+-   Added retrying in REST API requests.
+
 ## 0.1.0
 
 ### Minor Changes